Refactor Comment and Like controllers to use updated package structure and repository interfaces
