+++ src/main/java/com/springbootacademy/skill_sync/repo/CommentRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.springbootacademy.skill_sync.entity.Comment;
 
-public interface CommentRepo extends MongoRepository<Comment, String> {
+public interface CommentRepository extends MongoRepository<Comment, String> {
     List<Comment> findByPostId(String postId);
 }
